Added images as resources so they are included in jar

+++ src/tests/LoadAndSaveTests.java
@@ -202,7 +202,7 @@ public class LoadAndSaveTests {
 
         assert drawableRollingStock.getCurTrack().getId() == stockObject.getInt("curTrack");
 
-        assert stock.getLength() == stockObject.getInt("length");
+        assert stock.getLength() == stockObject.getDouble("length"); // TODO // FIXME: 8/10/16 Length in rolling stock
 
         assert drawableRollingStock.getOrientation() == stockObject.getBoolean("orientation");
